pyproject.toml
[project]
name = "django-ninja-starter"
version = "0.1.0"
description = "django-ninja-starter"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.20",
    "django-admin-interface>=0.29.4",
    "django-cors-headers>=4.7.0",
    "django-hint>=0.3.2",
    "django-ninja-extra>=0.22.6",
    "django-ninja-jwt>=5.3.5",
    "django-redis>=5.4.0",
    "ninja-schema>=0.14.1",
    "openpyxl>=3.1.5",
    "pyinstrument>=5.0.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]

[tool.ruff]
respect-gitignore = true
line-length = 88
indent-width = 4
output-format = "full"

# 常用目录文件夹忽略
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
ignore = ["F401"]  # 忽略的规则

per-file-ignores = { }

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

select = [# 选择的规则
    "F",
    "E",
    "W",
    "UP",
]

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
quote-style = "double"